requirements.txt
-i https://pypi.org/simple
annotated-types==0.7.0; python_version >= '3.8'
anyio==4.8.0; python_version >= '3.9'
click==8.1.8; python_version >= '3.7'
ecdsa==0.19.0; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
fastapi==0.115.8; python_version >= '3.8'
h11==0.14.0; python_version >= '3.7'
idna==3.10; python_version >= '3.6'
passlib==1.7.4
psycopg2-binary==2.9.10; python_version >= '3.8'
pyasn1==0.6.1; python_version >= '3.8'
pydantic==2.10.6; python_version >= '3.8'
pydantic-core==2.27.2; python_version >= '3.8'
pydantic-settings==2.7.1; python_version >= '3.8'
python-dotenv==1.0.1; python_version >= '3.8'
python-jose==3.3.0
pyyaml==6.0.2; python_version >= '3.8'
rsa==4.9; python_version >= '3.6' and python_version < '4'
six==1.17.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sniffio==1.3.1; python_version >= '3.7'
sqlalchemy==2.0.38; python_version >= '3.7'
starlette==0.45.3; python_version >= '3.9'
typing-extensions==4.12.2; python_version >= '3.8'
uvicorn==0.34.0; python_version >= '3.9'
